requirements.txt
bs4==0.0.2
fastapi==0.115.9
pandas==2.2.3
pydantic==2.10.6
requests==2.32.3
SQLAlchemy==2.0.38
urllib3==2.3.0
uvicorn==0.34.0
google-generativeai==0.8.4
PyMySQL==1.1.1
python-multipart==0.0.20
lxml==5.3.1